pyproject.toml
[tool.poetry]
name = "fit-tracker-backend"
version = "0.1.0"
description = ""
authors = ["Oleksii Martseniuk <martsenyk.l@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
django-ninja = "^1.1.0"
django-cors-headers = "^4.3.1"
django-ninja-extra = "^0.20.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"
isort = "^5.13.2"
pytest-django = "^4.8.0"
coverage = "^7.4.3"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
pythonpath = ". apps"
